+++ persistent-test/src/CompositeTest.hs
@@ -19,25 +19,10 @@
 module CompositeTest where
 
 import Test.Hspec.Expectations ()
-
-
-#  if MIN_VERSION_monad_control(0, 3, 0)
 import qualified Control.Monad.Trans.Control
-#  else
-import qualified Control.Monad.IO.Control
-#  endif
-
-#  if MIN_VERSION_monad_control(0, 3, 0)
 import qualified Control.Exception as E
-#    define CATCH catch'
-#  else
-import qualified Control.Exception.Control as Control
-#    define CATCH Control.catch
-#  endif
-
 import Init
 
-import Control.Applicative ((<$>),(<*>))
 #ifndef WITH_NOSQL
 import Data.Maybe (isJust)
 import Database.Persist.TH (mkDeleteCascade)
@@ -84,6 +69,13 @@ share [mkPersist persistSettings { mpsGeneric = False }, mkMigrate "compositeMig
     citizen CitizenId
     address AddressId
     Primary citizen address
+    deriving Eq Show
+
+  PrimaryCompositeWithOtherNullableFields
+    foo String       maxlen=20
+    bar String       maxlen=20
+    baz String Maybe
+    Primary foo bar
     deriving Eq Show
 |]
 
@@ -305,7 +297,6 @@ matchCitizenAddressK :: Key CitizenAddress -> Either Text (Int64, Int64)
 matchCitizenAddressK = (\(a:b:[]) -> (,) <$> fromPersistValue a <*> fromPersistValue b)
                      . keyToValues
 
-#if MIN_VERSION_monad_control(0, 3, 0)
 catch' :: (Control.Monad.Trans.Control.MonadBaseControl IO m, E.Exception e)
        => m a       -- ^ The computation to run
        -> (e -> m a) -- ^ Handler to invoke if an exception is raised
@@ -313,4 +304,3 @@ catch' :: (Control.Monad.Trans.Control.MonadBaseControl IO m, E.Exception e)
 catch' a handler = Control.Monad.Trans.Control.control $ \runInIO ->
                     E.catch (runInIO a)
                             (\e -> runInIO $ handler e)
-#endif