+++ persistent/Database/Persist/Sql/Orphan/PersistUnique.hs
@@ -13,11 +13,48 @@ import Database.Persist.Sql.Raw
 import Database.Persist.Sql.Orphan.PersistStore (withRawQuery)
 import Database.Persist.Sql.Util (dbColumns, parseEntityValues)
 import qualified Data.Text as T
-import Data.Monoid (mappend)
+import Data.Monoid (mappend, (<>))
 import qualified Data.Conduit.List as CL
 import Control.Monad.Trans.Reader (ask, withReaderT)
+import Control.Monad (when, liftM)
+
+defaultUpsert :: (MonadIO m, PersistEntity record, PersistUniqueWrite backend
+                 , PersistEntityBackend record ~ BaseBackend backend) 
+                => record -> [Update record] -> ReaderT backend m (Entity record)
+defaultUpsert record updates = do
+  uniqueKey <- onlyUnique record
+  upsertBy uniqueKey record updates
 
 instance PersistUniqueWrite SqlBackend where
+
+    upsert record updates = do
+      conn <- ask
+      uniqueKey <- onlyUnique record
+      case connUpsertSql conn of
+        Just upsertSql -> case updates of
+                            [] -> defaultUpsert record updates
+                            xs -> do
+                                let upds = T.intercalate "," $ map (go' . go) updates
+                                    sql = upsertSql t upds
+                                    vals = (map toPersistValue $ toPersistFields record) ++ (map updatePersistValue updates) ++ (unqs uniqueKey)
+                                           
+                                    go'' n Assign = n <> "=?"
+                                    go'' n Add = T.concat [n, "=", n, "+?"]
+                                    go'' n Subtract = T.concat [n, "=", n, "-?"]
+                                    go'' n Multiply = T.concat [n, "=", n, "*?"]
+                                    go'' n Divide = T.concat [n, "=", n, "/?"]
+                                    go'' _ (BackendSpecificUpdate up) = error $ T.unpack $ "BackendSpecificUpdate" `mappend` up `mappend` "not supported"
+                                              
+                                    go' (x, pu) = go'' (connEscapeName conn x) pu
+                                    go x = (fieldDB $ updateFieldDef x, updateUpdate x)
+
+                                x <- rawSql sql vals
+                                return $ head x
+        Nothing -> defaultUpsert record updates
+        where
+          t = entityDef $ Just record
+          unqs uniqueKey = concat $ map (persistUniqueToValues) [uniqueKey]
+
     deleteBy uniq = do
         conn <- ask
         let sql' = sql conn
@@ -69,3 +106,12 @@ instance PersistUniqueRead SqlWriteBackend where
 
 dummyFromUnique :: Unique v -> Maybe v
 dummyFromUnique _ = Nothing
+
+
+updateFieldDef :: PersistEntity v => Update v -> FieldDef
+updateFieldDef (Update f _ _) = persistFieldDef f
+updateFieldDef (BackendUpdate {}) = error "updateFieldDef did not expect BackendUpdate"
+
+updatePersistValue :: Update v -> PersistValue
+updatePersistValue (Update _ v _) = toPersistValue v
+updatePersistValue (BackendUpdate {}) = error "updatePersistValue did not expect BackendUpdate"