+++ persistent-mysql/ChangeLog.md
@@ -1,3 +1,9 @@
+## 2.6
+
+Compatibility for backend-specific upsert functionality.
+A lucky contributor could add upsert to the MySQL backend now, i.e.:
+INSERT ... ON DUPICATE ...
+
 ## 2.5
 
 * changes for read/write typeclass split